some fixes and nes SD card stuff

+++ Inc/main.h
@@ -170,7 +170,7 @@ typedef struct {
 	uint16_t MIN_TEMP;														/* Temperatur in K, UNTER der ein Interrupt ausgeloest wird */
 	uint16_t MAX_TEMP_RAW;												/* Oberer Schwellwert des Analog WDG*/
 	uint16_t MIN_TEMP_RAW;												/* Unterer Schwellwert des Analog WDG*/
-	uint16_t MAX_ACC;
+	float MAX_ACC;
 	uint8_t ACCELERATION_ENABLE;
 	uint8_t GNSS_ENABLE; 
 	uint8_t LIGHT_ENABLE;
